feat: set up basic Express server with CORS and body-parser middleware
