--- fakeApiBackend/backend/routes/product.js
@@ -1,12 +1,0 @@
-const express = require('express');
-const cors = require('cors');
-const bodyParser = require('body-parser');
-
-const app = express();
-
-app.use(cors());
-app.use(bodyParser.json());
-
-const PORT = 5000;
-
-app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
